requirements.txt
beautifulsoup4==4.7.1
docutils==0.14
emoji==0.5.4
Flask==1.1.1
Flask-Cors==3.0.8
future==0.17.1
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.6.0
gunicorn==20.0.4
h5py==2.9.0
html5lib==1.0.1
json5==0.8.4
jsonschema==3.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.0
more-itertools==7.0.0
nltk==3.4.4
numpy==1.16.4
numpydoc==0.9.1
oauthlib==3.1.0
pandas==0.24.2
py==1.8.0
python-dateutil==2.8.0
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.21.2
scipy==1.4.1
seaborn==0.9.0
tensorflow==2.1.0
tensorflow-datasets==1.3.2
tweepy==3.8.0
tweet-preprocessor==0.5.0